requirements.txt
amqp==5.3.1
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2
django-allauth==0.51.0
django-apscheduler==0.6.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
kombu==5.4.2
lxml==5.3.0
mpmath==1.3.0
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.2.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
sympy==1.13.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.8.0
wsproto==1.2.0
